Move get_interface_to_target() definition to LocalMachineInfo
This PR does two things:
    1. Define things close to where they're used.
    2. Remove vaguely named "tools.py"
